riot-control/coffee_break/requirements.txt
Django==1.9.4
django-pipeline==1.6.7
djangorestframework==3.3.3
